refactor: Move screens into same directory
